+++ src/assets/data/data.jsx
@@ -1,0 +1,36 @@
+const MOCK_FAMILY_MEMBERS = [
+  {
+    name: '김oo',
+    role: '아빠',
+    description: '아빠가 가끔 화장실 불을 안 끄고 나가신다.',
+    imgSrc:
+      'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcR1iHaiD-giZi3Jqu-1Pm-5RK-CQHisYtiwAQ&s',
+    isHost: false,
+  },
+  {
+    name: '조oo',
+    role: '엄마',
+    description: '엄마가 가장 좋아하는 꽃은 장미꽃이다.',
+    imgSrc:
+      'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcR1iHaiD-giZi3Jqu-1Pm-5RK-CQHisYtiwAQ&s',
+    isHost: false,
+  },
+  {
+    name: '김oo',
+    role: '여동생',
+    description: '',
+    imgSrc:
+      'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcR1iHaiD-giZi3Jqu-1Pm-5RK-CQHisYtiwAQ&s',
+    isHost: true,
+  },
+  {
+    name: '김oo',
+    role: '남동생',
+    description: '',
+    imgSrc:
+      'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcR1iHaiD-giZi3Jqu-1Pm-5RK-CQHisYtiwAQ&s',
+    isHost: false,
+  },
+];
+
+export default MOCK_FAMILY_MEMBERS;
