+++ src/components/common/alert/alert.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import * as S from './alert.style';
 import Exclamationmark from '../../../assets/images/Exclamationmark.png';
 
@@ -28,19 +27,6 @@ const Alert = ({ message, detailMessage, onConfirm, onCancel, isOpen }) => {
       </S.ModalContainer>
     </S.ModalBackground>
   );
-};
-
-Alert.propTypes = {
-  message: PropTypes.string.isRequired,
-  detailMessage: PropTypes.string,
-  onConfirm: PropTypes.func.isRequired,
-  onCancel: PropTypes.func,
-  isOpen: PropTypes.bool.isRequired,
-};
-
-Alert.defaultProps = {
-  detailMessage: null,
-  onCancel: null,
 };
 
 export default Alert;
