+++ src/constants/path.js
@@ -1,25 +1,27 @@
 const PAGE_PATH = {
-	// post
-	BASE: '/',
-	HOME: '/home',
-	// Login
-	LOGIN: 'login',
-	// Signup
-	SIGN_UP: 'signup',
-	// Service
-	SERVICE: 'service',
-	// Chat
-	CHAT: 'chat',
-	// Setting
-	SETTING: 'setting',
-	// Photo
-	PHOTO: 'photo',
-	// Calender
-	CALENDER: 'calender',
-	// Alert
-	ALERT: 'alert',
-	// Family
-	FAMILY: 'family',
+  // post
+  BASE: '/',
+  HOME: '/home',
+  // Login
+  LOGIN: 'login',
+  // Signup
+  SIGN_UP: 'signup',
+  // Service
+  SERVICE: 'service',
+  // Chat
+  CHAT: 'chat',
+  // Setting
+  SETTING: 'setting',
+  // Photo
+  PHOTO: 'photo',
+  // Calender
+  CALENDER: 'calender',
+  // Alert
+  ALERT: 'alert',
+  // Family
+  FAMILY: 'family',
+  // Family/Family-edit
+  FAMILY_EDIT: 'family-edit',
 };
 
 export { PAGE_PATH };
